Directly expose the async_task::Task and remove the non working spawn_local().
The spawn_local() did deadlock when used directly. So I removed it for now.
The main purpose for tasks is to run on the thread pool and then later be queries
by block_on and poll_once anyhow. Thread local doesn't make so much sense in that
context. It may make sense in the scoped executor, since we maybe want also to do
some work in the main thread while waiting for the other tasks to return.